--- src/app/page.module.css
@@ -1,18 +1,0 @@
-.header {
-    display: flex;
-    justify-content: space-between;
-    padding: 8px 16px;
-}
-
-.header ul {
-    display: flex;
-    list-style: none;
-    gap: 8px;
-}
-
-.header ul li {
-    padding: 8px;
-    border-radius: 4px;
-    background-color: #eee;
-    color: black;
-}
