+++ src/app/page.tsx
@@ -1,13 +1,5 @@
-'use client'
- 
-import { useRouter } from 'next/navigation'
- 
-export default function Page() {
-  const router = useRouter()
- 
-  return (
-    <button type="button" onClick={() => console.log('Hola')}>
-      Dashboard x
-    </button>
-  )
+import { redirect } from 'next/navigation';
+
+export default async function Home() {
+  redirect('/inicio');
 }
