+++ src/components/Header.tsx
@@ -1,0 +1,8 @@
+export default function Header() {
+    return (
+        <header>
+            <img src="\logo.png" alt="" />
+            <button typeof='button' >Cerrar Sesión</button>
+        </header>
+    )
+}
